Refactor to better categorize expressions
